Don't set IDs if reconstructing from database
This is a temporary hack that allows the create_workflow method to
assign (or not) IDs depending on how the method is being used. If
we are creating a Workflow from scratch, we want to set IDs. If we
are creating a Workflow from the databse version (reconstituting)
we don't need to set the IDs.

This is a temporary modification until we refactor the workflow
interface API.