Include Dockerfiles in the workflow tarball on submission
This fixes issue #551. Before, if the 'dockerFile' option was a relative path,
the TM would try to find the file based on where it was started. These changes
now require the dockerFile to be in the workflow tarball on submission, and
relative paths will refer to the tarred directory. This is done at parse time,
by reading in the dockerFile and storing it in the 'dockerFile' attribute.