Add 500 error handling to the bee client and the WFM
This adds an error handler in the WFM to catch exceptions and then forward
these on to the client. The `Connection` object now also has an optional error
handler. The BEE client uses this to dump the WFM exception, as well as the
arguments that caused this exception.

This only catches 500 errors, which isn't very user friendly. We'll need to add
some more error handling in both the WFM and the client as well.